% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_measurement.r
\name{analyze_measurement}
\alias{analyze_measurement}
\title{Analyze one measurment in binary validation}
\usage{
analyze_measurement(variable, dat, trans = "log2", clean = TRUE,
  contam = FALSE, prefix = "")
}
\arguments{
\item{variable}{String with the name of the variable to study. Must be a column
name in dat}

\item{dat}{data.frame with the measurments and metadata to analyze. Most contain columns:
PlateID, Col, Row, PicRep, Bacteria, Strain and Contaminated plus a column corresponding 
to variable. Other columns are ignored}

\item{trans}{String either "log2", "log2plus1" or "none".}

\item{clean}{logical indicating wheter observations of value zero (dead plants) should be removed}

\item{contam}{logical indicating whether to keep wells marked as contaminated}

\item{prefix}{string indicating prefix for created files}
}
\description{
Analyze one measurment in binary validation
}
\author{
Sur Herrera Paredes
}
